Move keeper methods to separate files
